13_computer_use_agent/requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
opencv-python>=4.8.0
torch>=2.1.0
transformers>=4.40.0
qwen-vl-utils>=0.0.8
accelerate>=0.30.0
pydantic>=2.0.0
